laying groundwork for simulation container refactor
