+++ eee/evolve/simulation/__init__.py
@@ -1,3 +1,3 @@
 
 from .load_json import load_json
-from .simulation_container import SimulationContainer
+from .container import SimulationContainer
