+++ example/test_run/ddg.csv
@@ -1,0 +1,77 @@
+,site,mut,hdna,h,l2e
+0,1,L1A,6.777999999999793,7.9409999999998035,4.438500000000204
+1,1,L1C,8.146999999999935,11.205000000000382,7.599999999999909
+2,1,L1D,-0.2339999999999236,1.07300000000032,3.291000000000168
+3,1,L1E,7.648000000000138,8.496000000000095,4.235999999999876
+4,1,L1F,1.4470000000001164,4.023000000000138,-0.4560000000001309
+5,1,L1G,-5.67699999999968,6.726999999999862,11.509999999999764
+6,1,L1H,4.436999999999898,2.6780000000003383,2.007999999999811
+7,1,L1I,5.106999999999971,13.08199999999988,3.2829999999999018
+8,1,L1K,5.624000000000251,3.4600000000000364,6.539999999999964
+10,1,L1M,2.2829999999999018,0.9769999999998618,0.5520000000001346
+11,1,L1N,7.119000000000142,4.653999999999996,3.7269999999998618
+12,1,L1P,1.5520000000001346,2.530999999999949,-19.751999999999956
+13,1,L1Q,5.0330000000003565,9.034000000000106,3.98299999999972
+14,1,L1R,7.016000000000076,0.9619999999999892,5.581000000000131
+15,1,L1S,0.3800000000001091,14.878999999999904,4.841999999999644
+16,1,L1T,4.983000000000175,0.5239999999998872,3.4589999999998327
+17,1,L1V,9.427999999999884,14.112999999999827,1.5790000000001785
+18,1,L1W,-2.708000000000084,2.341000000000349,-3.257000000000062
+19,1,L1Y,-0.8279999999999745,1.7660000000000764,-2.907000000000153
+20,2,L2A,21.615999999999985,7.046666666666472,12.66800000000012
+21,2,L2C,13.109000000000378,9.044000000000324,13.221000000000004
+22,2,L2D,16.635999999999967,12.803999999999633,15.751999999999953
+23,2,L2E,20.33199999999988,13.826000000000022,16.096999999999753
+24,2,L2F,-2.367999999999938,9.336999999999987,8.408000000000357
+25,2,L2G,15.92699999999968,-0.5650000000000546,12.784000000000106
+26,2,L2H,6.43100000000004,12.67599999999993,8.398999999999887
+27,2,L2I,2.2799999999997453,17.141999999999825,1.7229999999999563
+28,2,L2K,18.029,21.493000000000396,17.292999999999665
+30,2,L2M,8.233999999999924,19.56399999999985,12.484999999999673
+31,2,L2N,15.376999999999953,11.889000000000124,6.604000000000269
+32,2,L2P,-5.2970000000000255,38.60400000000027,44.48800000000029
+33,2,L2Q,14.775000000000093,5.832999999999629,10.66399999999976
+34,2,L2R,9.329000000000178,1.7260000000001128,14.26299999999992
+35,2,L2S,11.854499999999916,17.94699999999966,14.501999999999953
+36,2,L2T,12.907999999999902,5.248000000000047,6.964999999999691
+37,2,L2V,8.751999999999953,4.434999999999945,4.398000000000138
+38,2,L2W,9.518000000000027,18.83499999999958,16.208999999999833
+39,2,L2Y,-0.5779999999999745,6.382999999999811,8.28899999999976
+40,3,I3A,16.28099999999995,14.726000000000113,20.36700000000019
+41,3,I3C,14.500999999999747,12.625,18.840000000000146
+42,3,I3D,20.56799999999976,32.4409999999998,17.70199999999977
+43,3,I3E,18.58699999999999,18.677999999999884,23.11499999999978
+44,3,I3F,6.76299999999992,1.7219999999997526,5.396999999999935
+45,3,I3G,25.942000000000007,14.31899999999996,20.41699999999992
+46,3,I3H,11.956000000000133,11.327999999999976,14.29300000000012
+48,3,I3K,26.67700000000013,24.713000000000196,33.53099999999995
+49,3,I3L,3.3049999999998363,28.144000000000236,6.432999999999993
+50,3,I3M,3.925999999999931,6.170999999999822,17.529999999999745
+51,3,I3N,12.021999999999936,26.41800000000012,14.646999999999936
+52,3,I3P,25.56399999999985,36.93100000000004,38.58800000000019
+53,3,I3Q,14.993000000000391,14.56899999999996,27.057000000000244
+54,3,I3R,20.37599999999975,17.36999999999989,31.46199999999999
+55,3,I3S,19.269999999999985,12.006000000000313,18.772999999999683
+56,3,I3T,11.95699999999988,6.728999999999815,10.022999999999683
+57,3,I3V,4.572999999999865,4.129999999999654,3.7519999999999527
+58,3,I3W,9.564000000000306,34.6880000000001,16.817000000000007
+59,3,I3Y,9.763000000000376,7.375,7.3149999999996
+60,4,G4A,6.461999999999989,-9.967999999999847,19.19100000000026
+61,4,G4C,13.501999999999953,22.235000000000127,30.644999999999985
+62,4,G4D,22.911000000000055,46.05499999999984,37.54499999999962
+63,4,G4E,37.2170000000001,66.49699999999984,48.58400000000029
+64,4,G4F,48.047000000000025,63.46299999999974,29.34099999999989
+66,4,G4H,48.62900000000036,54.08800000000019,36.38200000000006
+67,4,G4I,50.14499999999998,35.1880000000001,28.062999999999647
+68,4,G4K,57.76699999999983,57.10599999999977,57.55000000000018
+69,4,G4L,35.35500000000002,35.87300000000005,29.92099999999982
+70,4,G4M,28.59500000000025,45.40499999999975,45.48000000000002
+71,4,G4N,15.47900000000027,36.297000000000025,30.126999999999956
+72,4,G4P,19.502000000000407,87.81799999999998,97.72700000000007
+73,4,G4Q,30.73400000000038,54.3449999999998,38.6279999999997
+74,4,G4R,33.409999999999854,90.31799999999998,56.6880000000001
+75,4,G4S,15.472000000000207,35.672999999999774,20.027999999999796
+76,4,G4T,15.427999999999884,43.396999999999935,36.32999999999993
+77,4,G4V,24.130999999999855,39.58999999999969,45.76800000000003
+78,4,G4W,79.75300000000016,96.20500000000015,46.90700000000015
+79,4,G4Y,52.32400000000007,80.4090000000001,33.11400000000003
