+++ example/test_run/simulation.json
@@ -1,0 +1,49 @@
+{
+  "population_size": 1000,
+  "mutation_rate": 0.01,
+  "num_generations": 1000,
+  "write_prefix": "eee_sim",
+  "write_frequency": 1000,
+  "seed": 874067739861,
+  "ens": {
+    "hdna": {
+      "dG0": 0.0,
+      "observable": true,
+      "mu_stoich": {},
+      "folded": true
+    },
+    "h": {
+      "dG0": 5.0,
+      "observable": false,
+      "mu_stoich": {},
+      "folded": true
+    },
+    "l2e": {
+      "dG0": 5.0,
+      "observable": false,
+      "mu_stoich": {
+        "iptg": 2.0
+      },
+      "folded": true
+    }
+  },
+  "R": 0.001987,
+  "mu_dict": {
+    "iptg": [
+      1.0,
+      4.0
+    ]
+  },
+  "select_on": "fx_obs",
+  "select_on_folded": true,
+  "fitness_kwargs": {},
+  "T": [
+    298.15,
+    298.15
+  ],
+  "fitness_fcns": [
+    "off",
+    "on"
+  ],
+  "ddg_df": "ddg.csv"
+}
